pyproject.toml
[tool.setuptools.packages.find]
include = ["llm_cls*"]
exclude = ["tests*", "configs*"]
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-cls"
version = "0.1.0"
description = "Decoder-based LLM classification with LoRA + 4-bit"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name="Amirhossein Yousefiramanddi" }]
dependencies = [
  "transformers>=4.41.0",
  "datasets>=2.19.0",
  "torchmetrics>=1.3.0",
  "peft>=0.11.1",
  "bitsandbytes>=0.43.0; platform_system != 'Windows'",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "pyyaml>=6.0.1",
  "typer>=0.12.0",
  "accelerate>=0.30.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.5.0", "pre-commit>=3.7.0"]

[tool.setuptools.package-data]
"llm_cls" = ["py.typed"]

[tool.ruff]
line-length = 110
target-version = "py310"
